backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
langchain==0.3.7
langchain-community==0.3.7
langchain-google-genai==2.0.6
langchain-openai==0.2.9
langchain-huggingface==0.1.1
langchain-groq==0.2.1
langchain-aws>=0.2.24
boto3>=1.39.7
sentence-transformers==3.3.1
langchain-text-splitters==0.3.2
chromadb==0.5.20
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.3
python-dotenv==1.0.1
tiktoken==0.8.0
google-generativeai>=0.8.0
pillow>=10.0.0
voyageai>=0.3.0
pinecone>=5.0.0
supabase>=2.0.0
